% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cv.SplitReg}
\alias{coef.cv.SplitReg}
\title{Extract coefficients from a cv.SplitReg object.}
\usage{
\method{coef}{cv.SplitReg}(object, index = object$index_opt, ...)
}
\arguments{
\item{object}{Fitted cv.SplitReg object.}

\item{index}{Indices indicating values of lambda_S at which to extract coefficients. Defaults to the optimal value.}

\item{...}{Additional arguments for compatibility}
}
\value{
A vector of coefficients
}
\description{
Extract coefficients from a cv.SplitReg object.
}
\seealso{
\code{\link{cv.SplitReg}}
}
