% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitReg_methods_R.R
\name{coef.SplitReg}
\alias{coef.SplitReg}
\title{Extract coefficients from a SplitReg object.}
\usage{
\method{coef}{SplitReg}(object, ...)
}
\arguments{
\item{object}{Fitted cv.SplitReg object.}

\item{...}{Additional arguments for compatibility}
}
\value{
A vector of coefficients
}
\description{
Extract coefficients from a SplitReg object.
}
\examples{
library(MASS)
set.seed(1)
beta <- c(rep(5, 5), rep(0, 45))
Sigma <- matrix(0.5, 50, 50)
diag(Sigma) <- 1
x <- mvrnorm(50, mu = rep(0, 50), Sigma = Sigma)
y <- x \%*\% beta + rnorm(50)
fit <- SplitReg(x, y, num_models=2)
split.coefs <- coef(fit)

}
\seealso{
\code{\link{SplitReg}}
}
